producerVelib/requirements.txt
DateTime==4.3
requests==2.25.0
urllib3==1.26.1
kafka-python==2.0.2
